+++ test/queryCourses.test.js
@@ -3,7 +3,10 @@ import queryCourses from '../src/queryCourses'
 
 describe('queryCourses', () => {
 	it('queries a list of courses', () => {
-		let query = 'dept: AMCON year: 2013'
+		let query = {
+			depts: ['AMCON'],
+			year: [2013],
+		}
 		let courses = [
 			{depts: ['AMCON'], year: 2013}
 		]
@@ -11,7 +14,9 @@ describe('queryCourses', () => {
 	})
 
 	it('properly handles a list of five years', () => {
-		let query = 'year: $OR year: 2010 year: 2011 year: 2012 year: 2013 year: 2014'
+		let query = {
+			year: ['$OR', 2010, 2011, 2012, 2013, 2014]
+		}
 		let courses = [
 			{depts: ['AMCON'], year: 2013},
 			{depts: ['ASIAN'], year: 2012},