+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Updated 6to5 to v3.0.0
 - Various internal optimizations
 - **Removed:** `getRandomInt`
+- **Changed:** `queryCourses` no longer accepts a query string; it now expects a query object, like the one from `buildQueryFromString`
 - **Removed:** `isAsianCon` has moved back into `sto-areas`
 - **Changed:** `combinations` now uses generators by default. The old behavior is available under `basicCombinations`